requirements.txt
python-telegram-bot[job-queue]==20.7
python-dotenv==1.0.0
SQLAlchemy==2.0.23
APScheduler>=3.6.3
pymorphy2==0.9.1
